app.py
import logging
import os
import sqlite3
import tempfile
from datetime import datetime

from flask import Flask, jsonify, render_template, request
from model import get_emotion_detector, predict_emotion
from PIL import Image
from werkzeug.utils import secure_filename

# Configure logging for production
logging.basicConfig(
    level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
)
logger = logging.getLogger(__name__)

app = Flask(__name__)
app.config["SECRET_KEY"] = os.environ.get(
    "SECRET_KEY", "emotion_detection_secret_key_2024"
)
app.config["MAX_CONTENT_LENGTH"] = 16 * 1024 * 1024

ALLOWED_EXTENSIONS = {"png", "jpg", "jpeg", "gif", "bmp", "tiff", "webp"}

# Create uploads directory
UPLOAD_FOLDER = "uploads"
os.makedirs(UPLOAD_FOLDER, exist_ok=True)


def allowed_file(filename):
    return "." in filename and filename.rsplit(".", 1)[1].lower() in ALLOWED_EXTENSIONS


def get_database_stats():
    try:
        conn = sqlite3.connect("emotion_predictions.db")
        cursor = conn.cursor()

        cursor.execute("SELECT COUNT(*) FROM predictions")
        total_predictions = cursor.fetchone()[0]

        cursor.execute(
            "SELECT predicted_emotion, COUNT(*) FROM predictions GROUP BY predicted_emotion ORDER BY COUNT(*) DESC"
        )
        emotion_counts = dict(cursor.fetchall())

        cursor.execute(
            "SELECT timestamp, predicted_emotion, confidence, source FROM predictions ORDER BY timestamp DESC LIMIT 5"
        )
        recent_predictions = cursor.fetchall()

        conn.close()

        return {
            "total_predictions": total_predictions,
            "emotion_counts": emotion_counts,
            "recent_predictions": recent_predictions,
        }
    except Exception as e:
        logger.error(f"Error getting database stats: {str(e)}")
        return {"total_predictions": 0, "emotion_counts": {}, "recent_predictions": []}


@app.route("/")
def index():
    try:
        stats = get_database_stats()
        return render_template(
            "index.html", stats=stats, title="Emotion Detection Web App"
        )
    except Exception as e:
        logger.error(f"Error in index route: {str(e)}")
        return render_template(
            "index.html",
            stats={
                "total_predictions": 0,
                "emotion_counts": {},
                "recent_predictions": [],
            },
            title="Emotion Detection Web App",
        )


@app.route("/predict", methods=["POST"])
def predict():
    try:
        if "file" not in request.files:
            return jsonify({"error": "No file uploaded"}), 400

        file = request.files["file"]

        if file.filename == "":
            return jsonify({"error": "No file selected"}), 400

        if not allowed_file(file.filename):
            return jsonify(
                {"error": "File type not allowed. Please upload an image file."}
            ), 400

        if file and allowed_file(file.filename):
            try:
                with tempfile.NamedTemporaryFile(
                    delete=False, suffix=".jpg"
                ) as temp_file:
                    file.save(temp_file.name)

                    try:
                        img = Image.open(temp_file.name)
                        img.verify()
                        img = Image.open(temp_file.name)

                        if img.mode != "RGB":
                            img = img.convert("RGB")

                    except Exception as img_error:
                        os.unlink(temp_file.name)
                        return jsonify(
                            {"error": f"Invalid image file: {str(img_error)}"}
                        ), 400

                    logger.info(
                        f"Making prediction for uploaded image: {file.filename}"
                    )
                    result = predict_emotion(temp_file.name, source="flask")

                    os.unlink(temp_file.name)

                    response_data = {
                        "success": True,
                        "predicted_emotion": result["emotion"],
                        "confidence": round(result["confidence"], 3),
                        "all_scores": {
                            k: round(v, 3) for k, v in result["all_scores"].items()
                        },
                        "filename": secure_filename(file.filename),
                        "timestamp": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
                    }

                    logger.info(
                        f"Prediction successful: {result['emotion']} ({result['confidence']:.3f})"
                    )
                    return jsonify(response_data)

            except Exception as process_error:
                logger.error(f"Error processing image: {str(process_error)}")
                return jsonify(
                    {"error": f"Error processing image: {str(process_error)}"}
                ), 500

    except Exception as e:
        logger.error(f"Error in predict route: {str(e)}")
        return jsonify({"error": f"Server error: {str(e)}"}), 500


@app.route("/health")
def health_check():
    try:
        conn = sqlite3.connect("emotion_predictions.db")
        cursor = conn.cursor()
        cursor.execute("SELECT COUNT(*) FROM predictions")
        count = cursor.fetchone()[0]
        conn.close()

        detector = get_emotion_detector()
        model_loaded = detector.model is not None

        return jsonify(
            {
                "status": "healthy",
                "model_loaded": model_loaded,
                "database_predictions": count,
                "timestamp": datetime.now().isoformat(),
            }
        )

    except Exception as e:
        return jsonify(
            {
                "status": "unhealthy",
                "error": str(e),
                "timestamp": datetime.now().isoformat(),
            }
        ), 500


@app.route("/stats")
def stats():
    try:
        stats = get_database_stats()
        return jsonify(stats)
    except Exception as e:
        logger.error(f"Error getting stats: {str(e)}")
        return jsonify({"error": str(e)}), 500


@app.errorhandler(413)
def too_large(e):
    return jsonify({"error": "File too large. Maximum size is 16MB."}), 413


@app.errorhandler(500)
def internal_error(error):
    logger.error(f"Internal server error: {str(error)}")
    return jsonify({"error": "Internal server error"}), 500


@app.errorhandler(404)
def not_found(error):
    return jsonify({"error": "Not found"}), 404


# Initialize model on startup
try:
    logger.info("Initializing emotion detection model...")
    detector = get_emotion_detector()
    logger.info("Model initialized successfully!")
except Exception as e:
    logger.error(f"Error initializing model: {str(e)}")
    # Continue running even if model fails to load initially


if __name__ == "__main__":
    # Use environment variables for deployment
    port = int(os.environ.get("PORT", 5000))
    debug = os.environ.get("FLASK_ENV") != "production"

    logger.info(f"Starting application on port {port}")
    logger.info(f"Debug mode: {debug}")
    logger.info(f"Environment: {os.environ.get('FLASK_ENV', 'development')}")

    app.run(host="0.0.0.0", port=port, debug=debug)
