requirements.txt
# Core web framework
flask==2.3.3
Werkzeug==2.3.7

# Production server
gunicorn==21.2.0

# Machine learning and AI
torch>=2.0.0,<2.2.0
transformers>=4.30.0,<4.40.0
huggingface-hub>=0.16.0,<0.20.0

# Image processing
Pillow>=9.0.0,<11.0.0

# Data handling
numpy>=1.24.0,<1.26.0
requests>=2.28.0,<2.32.0

# Optional for better performance
accelerate>=0.20.0,<0.25.0

# Model saving and loading
h5py>=3.8.0,<3.11.0

# Security and file handling
cryptography>=3.4.8

# Development and debugging
python-dotenv>=1.0.0
